Implement Timer-A emulation (delay-mode only). Needed by NetBSD.

+++ src/Unix/host_clock_unix.cpp
@@ -25,10 +25,10 @@
 
 /*--- Private stuff ---*/
 
-float HostClockUnix::getCurTime(void)
+uint32 HostClockUnix::getCurTime(void)
 {
 	struct timeval tv;
 	gettimeofday(&tv, NULL);
 
-	return ((tv.tv_sec * 1000.0) + (tv.tv_usec / 1000.0));
+	return ((tv.tv_sec * 1000) + (tv.tv_usec / 1000));
 }
